+++ quiz_quest/lib/app/features/quiz_pages/general_quiz_pages/hard_general_quiz_page/hard_question_quiz_page.dart
@@ -20,7 +20,9 @@ import 'package:quiz_quest/app/features/quiz_pages/games_quiz_pages/easy_games_q
 import 'package:quiz_quest/app/features/quiz_pages/general_quiz_pages/cubit/general_cubit.dart';
 import 'package:quiz_quest/app/features/quiz_pages/general_quiz_pages/easy_general_quiz_page/easy_general_lost_life_page.dart';
 import 'package:quiz_quest/app/features/quiz_pages/general_quiz_pages/easy_general_quiz_page/resume_easy_general_question_quiz_page.dart';
+import 'package:quiz_quest/app/features/quiz_pages/general_quiz_pages/hard_general_quiz_page/hard_general_answer_button.dart';
 import 'package:quiz_quest/app/features/quiz_pages/general_quiz_pages/hard_general_quiz_page/hard_general_lost_life_page.dart';
+import 'package:quiz_quest/app/features/quiz_pages/general_quiz_pages/hard_general_quiz_page/hard_general_question_widget.dart';
 import 'package:quiz_quest/app/features/quiz_pages/general_quiz_pages/hard_general_quiz_page/resume_hard_general_question_quiz_page.dart';
 import 'package:quiz_quest/app/features/quiz_pages/geography_quiz_pages/cubit/geography_cubit.dart';
 import 'package:quiz_quest/app/features/quiz_pages/geography_quiz_pages/easy_geography_quiz_page/easy_geography_lost_life_page.dart';
@@ -281,7 +283,7 @@ class _HardQuestionGeneralQuizPageState
                             return Column(
                               mainAxisAlignment: MainAxisAlignment.spaceAround,
                               children: [
-                                QuestionWidget(
+                                HardGeneralQuestionWidget(
                                   question: generalQuizModel
                                       .results[currentIndex].question,
                                 ),
@@ -327,7 +329,7 @@ class _HardQuestionGeneralQuizPageState
                                 for (int index = 0;
                                     index < currentAnswers.length;
                                     index++) ...[
-                                  AnswerButton(
+                                  HardGeneralAnswerButton(
                                     isTimeUp: isTimeUp,
                                     duration: duration,
                                     isButtonDisabled: (value) {
@@ -455,150 +457,5 @@ class _HardQuestionGeneralQuizPageState
   }
 }
 
-class QuestionWidget extends StatelessWidget {
-  const QuestionWidget({
-    required this.question,
-    super.key,
-  });
 
-  final String question;
 
-  @override
-  Widget build(BuildContext context) {
-    final modifiedQuestion = question
-        .replaceAll('&quot;', '')
-        .replaceAll('&#039;', '')
-        .replaceAll('&aacute;', '')
-        .replaceAll('&ntilde;', '')
-        .replaceAll('&amp;', '')
-        .replaceAll('&rsquo;', '');
-
-    return Center(
-      child: Text(
-        modifiedQuestion,
-        style: GoogleFonts.aBeeZee(
-            fontSize: 24, color: Colors.white, fontWeight: FontWeight.bold),
-        textAlign: TextAlign.center,
-      ),
-    );
-  }
-}
-
-class AnswerButton extends StatefulWidget {
-  AnswerButton({
-    required this.answer,
-    required this.controller,
-    required this.isCorrectAnswer,
-    required this.colorFunction,
-    required this.isButtonClicked,
-    required this.isButtonDisabled,
-    required this.textcolor,
-    required this.index,
-    required this.duration,
-    required this.isTimeUp,
-  });
-
-  final String answer;
-  final CountDownController controller;
-  final bool isCorrectAnswer;
-
-  int duration;
-  bool isTimeUp;
-  Function(Color, int) colorFunction;
-  final Function(bool) isButtonClicked;
-  final Function(bool) isButtonDisabled;
-  Color textcolor;
-  final int index;
-
-  @override
-  State<AnswerButton> createState() => _AnswerButtonState();
-}
-
-class _AnswerButtonState extends State<AnswerButton> {
-  void onPressed() {
-    if (isButtonDisabled) {
-      return;
-    }
-
-    widget.controller.pause();
-
-    if (widget.isCorrectAnswer) {
-      widget.colorFunction(Colors.green, widget.index);
-      hardGeneralGoodAnswers += 1;
-    } else {
-      widget.colorFunction(Colors.red, widget.index);
-      hardGeneralBadAnswers += 1;
-      if (hardGeneralBadAnswers == 3) {
-        context
-            .read<GeneralCubit>()
-            .updateHardGeneralPoints(hardGeneralGoodAnswers);
-        Navigator.of(context).push(
-          MaterialPageRoute(
-            builder: (context) =>
-                HardGeneralLostLifePage(goodAnswers: hardGeneralGoodAnswers),
-          ),
-        );
-      }
-    }
-    setState(() {
-      widget.textcolor = widget.isCorrectAnswer ? Colors.green : Colors.red;
-    });
-    widget.isButtonClicked(true);
-    widget.isButtonDisabled(true);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final modifiedAnswer = widget.answer
-        .replaceAll('&quot;', '')
-        .replaceAll('&#039;', '')
-        .replaceAll('&aacute;', '')
-        .replaceAll('&ntilde;', '')
-        .replaceAll('&amp;', '')
-        .replaceAll('&rsquo;', '')
-        .replaceAll('&lrm;', '')
-        .replaceAll('&Eacute;', '');
-
-    return Container(
-      margin: const EdgeInsets.symmetric(horizontal: 20),
-      decoration: BoxDecoration(
-        gradient: const LinearGradient(
-          colors: [
-            Color.fromRGBO(11, 22, 65, 1),
-            Color.fromRGBO(9, 77, 203, 1),
-          ],
-          begin: Alignment.centerLeft,
-          end: Alignment.centerRight,
-        ),
-        borderRadius: const BorderRadius.all(
-          Radius.circular(6.0),
-        ),
-        boxShadow: [
-          BoxShadow(
-            color: Colors.black.withOpacity(0.1),
-            spreadRadius: 4,
-            blurRadius: 10,
-            offset: const Offset(0, 3),
-          )
-        ],
-      ),
-      child: ElevatedButton(
-        onPressed: onPressed,
-        style: ElevatedButton.styleFrom(
-          minimumSize: const Size.fromHeight(50),
-          backgroundColor: Colors.transparent,
-          shadowColor: Colors.transparent,
-        ),
-        child: Text(
-          modifiedAnswer,
-          style: GoogleFonts.aBeeZee(
-            fontSize: 24,
-            color: widget.isTimeUp && widget.isCorrectAnswer
-                ? Colors.green
-                : widget.textcolor,
-          ),
-        ),
-      ),
-    );
-  }
-}
